refactor: 프로필 공개 데이터 조회 캐시 적용
