+++ lib/cacheTags.ts
@@ -7,4 +7,9 @@ export const CACHE_TAGS = {
     userRankings: "user-rankings",
     exams: "exams",
     bingos: "bingos",
+    userProfiles: "user-profiles",
 } as const;
+
+export function getUserProfileTag(userId: number) {
+    return `user-profile-${userId}`;
+}
